Organize plugin modules
Plugin modules are now organized into directories.

This certainly fixes the issue of having a config that's hard to look
at, but the way in which this is done still feels very off. I will
have to look into how to do this in a much cleaner fashion some other
time.